Fix month calendar's events positioning
The js did not count the correct number of days to place de event. It just walk N days according the event's day number.
If the calendar show days from the previous month, the events were positionings on the wrong day.

Now, its calculate the amount of day from the first date on the calendar to the events date and place it on the right day.